Update package.json
Thanks for the plugin!

I want to use your plugin in my project where we use node 16.x. I get an error about incompatible node versions when installing the plugin.
I verified it works with 14.x, 16.x.

Would it be possible to change it so that we can install it without the incompatible engine error?